An attempt to resolve #8. Take declaration contexts into account.
The fundamental problem of not knowing exactly what this rule wants to prohibit
and what not still remains. Does it apply to all identifiers regardless of
context and scopes? Does it only apply to identifiers of the same kind? Does it
dictate that "namespace dc" clashes with "DeclContext *DC" even though mixing
them up is pretty much impossible?

Let's see if this commit is a usable interpretation of it. Improvements welcome!